Split up sklearn testing for optimization with bias minimization
